pyproject.toml
[project]
name = "mysolid-homeassistant"
version = "0.1.0"
description = "Home Assistant custom integration for MySolid / Solid Security."
readme = "README.md"
requires-python = ">=3.11"

[dependency-groups]
dev = [
    "aiohttp>=3.13.3",
    "cryptography>=46.0.5",
    "PyTurboJPEG>=1.8.2 ; python_full_version >= '3.13.2'",
    "protobuf==6.31.1",
    "pytest-homeassistant-custom-component==0.13.316 ; python_full_version >= '3.13.2'",
    "ruff>=0.13.0",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "--import-mode=importlib"
testpaths = ["tests"]
pythonpath = ["."]
filterwarnings = []

[tool.ruff]
line-length = 88
target-version = "py311"
extend-exclude = ["custom_components/mysolid/_proto"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]
